+++ languages/codeql/runnables.scm
@@ -1,0 +1,4 @@
+(
+  (select) @run
+  (#set! tag codeql-query)
+)